+++ baybe/utils/sampling_algorithms.py
@@ -12,10 +12,22 @@ import pandas as pd
 from baybe.utils.basic import is_all_instance
 
 
+class FPSInitialization(Enum):
+    """Initialization methods for farthest point sampling."""
+
+    FARTHEST = "farthest"
+    """Selects the first two points with the largest distance."""
+
+    RANDOM = "random"
+    """Selects the first point uniformly at random."""
+
+
 def farthest_point_sampling(
     points: np.ndarray,
     n_samples: int = 1,
-    initialization: Literal["farthest", "random"] | Collection[int] = "farthest",
+    initialization: FPSInitialization
+    | Literal["farthest", "random"]
+    | Collection[int] = "farthest",
     random_tie_break: bool = True,
 ) -> list[int]:
     """Select a subset of points using farthest point sampling.
@@ -96,11 +108,14 @@ def farthest_point_sampling(
                 f"range of available points (0 to {n_points - 1}) but contains "
                 f"out-of-bounds indices: {problematic_indices}"
             )
-    elif initialization not in {"farthest", "random"}:
-        raise ValueError(
-            f"Unknown initialization type. Expected 'farthest', 'random', or a "
-            f"collection of integers. Provided: {initialization=}"
-        )
+    else:
+        try:
+            initialization = FPSInitialization(initialization)
+        except ValueError as ex:
+            raise ValueError(
+                f"Unknown initialization type. Expected 'farthest', 'random', or a "
+                f"collection of integers. Provided: {initialization=}"
+            ) from ex
 
     if n_samples > n_points:
         raise ValueError(
@@ -126,18 +141,16 @@ def farthest_point_sampling(
     np.fill_diagonal(dist_matrix, -np.inf)
 
     # Initialize the point selection
-    if initialization == "random":
+    if initialization is FPSInitialization.RANDOM:
         selected_point_indices = [np.random.randint(0, n_points)]
-    elif initialization == "farthest":
+    elif initialization is FPSInitialization.FARTHEST:
         idx_1d = np.argmax(dist_matrix)
         selected_point_indices = list(
             map(int, np.unravel_index(idx_1d, dist_matrix.shape))
         )
         if n_samples == 1:
             return [sort_idx[selected_point_indices[0]]]
-    elif isinstance(initialization, Collection) and is_all_instance(
-        initialization, int
-    ):
+    else:  # initialization is Collection[int]
         inv_sort_idx = np.argsort(sort_idx)
         selected_point_indices = [inv_sort_idx[x] for x in initialization]