+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   are now keyword-only
 - `df_apply_permutation_augmentation` has a different interface and now expects
   permutation groups instead of column groups
+- `ParameterSelectorProtocol.__call__` now declares its input as positional-only
 
 ### Fixed
 - `DiscretePermutationInvarianceConstraint` no longer erroneously removes points where